Safari-609.1.11.1/LayoutTests/editing/pasteboard/merge-end-table-2-expected.txt

EDITING DELEGATE: shouldBeginEditingInDOMRange:range from 0 of DIV > BODY > HTML > #document to 1 of DIV > BODY > HTML > #document
EDITING DELEGATE: webViewDidBeginEditing:WebViewDidBeginEditingNotification
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: shouldChangeSelectedDOMRange:range from 7 of #text > DIV > TD > TR > TBODY > TABLE > DIV > BODY > HTML > #document to 7 of #text > DIV > TD > TR > TBODY > TABLE > DIV > BODY > HTML > #document toDOMRange:range from 3 of #text > DIV > TD > TR > TBODY > TABLE > DIV > BODY > HTML > #document to 3 of #text > DIV > TD > TR > TBODY > TABLE > DIV > BODY > HTML > #document affinity:NSSelectionAffinityDownstream stillSelecting:FALSE
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: webViewDidChange:WebViewDidChangeNotification
EDITING DELEGATE: webViewDidEndEditing:WebViewDidEndEditingNotification
This tests pasting text into a table cell.  The last bit of content in the incoming fragment should be merged with the paragraph after the position being pasted into.

Before insertion:
| <tbody>
|   <tr>
|     <td>
|       <div>
|         id="test2"
|         "Item 1"
|     <td>
|       "Item 2"

After insertion:
| <tbody>
|   <tr>
|     <td>
|       <div>
|         id="test2"
|         "Itemfoo"
|       <div>
|         id="test2"
|         "bar<#selection-caret> 1"
|     <td>
|       "Item 2"
